fix(send-slack-notification): Provide failed-tests default value
This condition is encountered if no integration test ran, but the
step still failed. This can be the case if other steps failed, like
creating the cluster, applying other resources to said cluster, etc.
The field in the Slack message needs to be non-empty. As such, we
now provide a default value.