chore: Comment out empty attributes processor
As per https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/c3e7329aeb9e6f273e54fda09c43719c205b14d9/processor/attributesprocessor/config.go#L35
an empty list of actions is reported as if the key is missing entirely.
This is honestly a weird decision, but we have to deal with it. For
now, we just comment out, because we don't actually list any attributes
we want to attach to logs.