feat(publish-image*): Add cosign retries (#83)
We are seeing infrequent cosign failures. Most of them are caused
by external the Fulcio service being unavailable for short bursts
of time. To make the publish-image and publish-image-index-manifest
actions more resilient against this flakiness, this PR introduces
a generic retry script which will run the specified command N times
with a timeout between individual retries.

By default, the cosign commands are retried 3 times with a timeout
of 30 seconds. These values can be overridden by the the following
action inputs:

- cosign-retries
- cosign-retry-timeout